EOD WIP: most of the code is broadly where it will end up, but this doesn't even think of building yet.
