+++ papi/logging.go
@@ -1,4 +1,4 @@
-package pto3
+package papi
 
 import (
 	"fmt"
