+++ papi/pagination.go
@@ -4,4 +4,4 @@
 // (1) take a request and a default page size and return a range, and
 // (2) take a JSON chunk ? and fill in appropriate next: and prev: links.
 
-package pto3
+package papi
